Fixes circular dependencies
