+++ packages/app-store/_appRegistry.ts
@@ -1,14 +1,14 @@
 import { App } from "@calcom/types/App";
 
-import appStore from ".";
+import appStoreMetadata from "./metadata";
 
 /** Mainly to use in listings for the frontend, use in getStaticProps or getServerSideProps */
 export function getAppRegistry() {
-  return Object.values(appStore).reduce((apps, app) => {
+  return Object.values(appStoreMetadata).reduce((apps, app) => {
     // Skip if app isn't installed
-    if (!app.metadata.installed) return apps;
+    if (!app.installed) return apps;
     // Let's not leak api keys to the front end
-    const { key, ...metadata } = app.metadata;
+    const { key, ...metadata } = app;
     apps.push(metadata);
     return apps;
   }, [] as Omit<App, "key">[]);
