+++ packages/app-store/metadata.ts
@@ -1,0 +1,33 @@
+import { metadata as applecalendar } from "./applecalendar/_metadata";
+import { metadata as caldavcalendar } from "./caldavcalendar/_metadata";
+import { metadata as dailyvideo } from "./dailyvideo/_metadata";
+import { metadata as googlecalendar } from "./googlecalendar/_metadata";
+import { metadata as googlevideo } from "./googlevideo/_metadata";
+import { metadata as huddle01video } from "./huddle01video/_metadata";
+import { metadata as jitsivideo } from "./jitsivideo/_metadata";
+import { metadata as office365calendar } from "./office365calendar/_metadata";
+import { metadata as office365video } from "./office365video/_metadata";
+import { metadata as slackmessaging } from "./slackmessaging/_metadata";
+import { metadata as stripepayment } from "./stripepayment/_metadata";
+import { metadata as tandemvideo } from "./tandemvideo/_metadata";
+import { metadata as wipemycalother } from "./wipemycalother/_metadata";
+import { metadata as zoomvideo } from "./zoomvideo/_metadata";
+
+export const appStoreMetadata = {
+  applecalendar,
+  caldavcalendar,
+  dailyvideo,
+  googlecalendar,
+  googlevideo,
+  huddle01video,
+  jitsivideo,
+  office365calendar,
+  office365video,
+  slackmessaging,
+  stripepayment,
+  tandemvideo,
+  zoomvideo,
+  wipemycalother,
+};
+
+export default appStoreMetadata;
